+++ src/LionWeb.Protocol.Delta/DeltaCommandToDeltaEventMapper.cs
@@ -24,6 +24,10 @@ using Core.Serialization;
 using Message.Command;
 using Message.Event;
 
+/// <summary>
+/// This class is used only in tests, will be moved to tests.
+/// </summary>
+[Obsolete("Not useful in production, will be removed.")]
 public class DeltaCommandToDeltaEventMapper
 {
     private readonly SharedNodeMap _sharedNodeMap;
@@ -58,7 +62,8 @@ public class DeltaCommandToDeltaEventMapper
             DeleteReference a => new ReferenceDeleted(a.Parent, a.Reference, a.Index, a.DeletedTarget, a.DeletedResolveInfo, OriginCommands(a), []),
             ChangeReference a => new ReferenceChanged(a.Parent, a.Reference, a.Index, a.NewTarget, a.NewResolveInfo, a.OldTarget, a.OldResolveInfo, OriginCommands(a), []),
             AddPartition a => new PartitionAdded(a.NewPartition, /*TODO*/ null, OriginCommands(a), []),
-            DeletePartition a => new PartitionDeleted(a.DeletedPartition, /*TODO*/[], OriginCommands(a), [])
+            DeletePartition a => new PartitionDeleted(a.DeletedPartition, /*TODO*/[], OriginCommands(a), []),
+            _ => throw new ArgumentException($"{nameof(DeltaCommandToDeltaEventMapper)} does not support {deltaCommand.GetType().Name}!")
         };
 
     private NodeId GetParent(NodeId childId)