+++ test/LionWeb.Core.Test/Notification/Replicator/ForestTests.cs
@@ -117,7 +117,7 @@ public class ForestTests : ReplicatorTestsBase
 
         originalForest.AddPartitions([destinationPartition]);
         // Node "moved" is unknown in the cloned forest
-        Assert.ThrowsExactly<NotImplementedException>(() => destinationPartition.AddLinks([moved]));
+        Assert.ThrowsExactly<InvalidOperationException>(() => destinationPartition.AddLinks([moved]));
     }
 
     [TestMethod]